showPowerSettingss condition change

+++ vtr-ui/src/app/components/pages/page-device-settings/children/subpage-device-settings-power/subpage-device-settings-power.component.ts
@@ -1090,7 +1090,7 @@ export class SubpageDeviceSettingsPowerComponent implements OnInit, OnDestroy {
 	}
 
 	showPowerSettings() {
-		if (this.isDesktopMachine || (!this.showEasyResumeSection && !this.alwaysOnUSBStatus.available) || this.showFlipToBootSection$.value) {
+		if (this.isDesktopMachine || (!this.showEasyResumeSection && !this.alwaysOnUSBStatus.available && !this.showFlipToBootSection$.value)) {
 			this.headerMenuItems = this.commonService.removeObjFrom(this.headerMenuItems, 'power');
 			// return false;
 		}
